rack/middleware: factor out Rails subscriber to a separate file
I hadn't known that Rails hooks can accept classes. This new way provides better
separation of concerns, so the middleware class is not polluted with route
stats stuff.
