+++ lib/airbrake/rails.rb
@@ -5,6 +5,8 @@ module Airbrake
     # occurring in the application automatically.
     class Railtie < ::Rails::Railtie
       initializer('airbrake.middleware') do |app|
+        require 'airbrake/rails/action_controller_subscriber'
+
         # Since Rails 3.2 the ActionDispatch::DebugExceptions middleware is
         # responsible for logging exceptions and showing a debugging page in
         # case the request is local. We want to insert our middleware after
